feat: add layer of security for deletion of records

+++ app/Http/Controllers/GeoInfoController.php
@@ -5,16 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\GeoInfo;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use App\Http\Requests\GeoInfoFormRequest;
-use App\Http\Resources\GeoInfoResource;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Services\GeoInfoService;
+use App\Http\Resources\GeoInfoResource;
+use App\Http\Requests\GeoInfoFormRequest;
 
 class GeoInfoController extends Controller
 {
-    public function index()
-    {
-
-    }
+    public function index() {}
 
     public function show(GeoInfo $geoInfo)
     {
@@ -24,24 +22,28 @@ class GeoInfoController extends Controller
     public function store(
         GeoInfoFormRequest $request,
         GeoInfoService $service
-    ){
+    ) {
         return DB::transaction(function () use ($request, $service) {
             $result = $service->make($request);
             return response()->json(new GeoInfoResource($result));
         });
     }
 
-    public function update()
-    {
-
-    }
+    public function update() {}
 
     public function destroy(GeoInfo $geoInfo)
     {
-        $geoInfo->delete();
+        if ($geoInfo->user_id == Auth::user()->id) {
+            $geoInfo->delete();
+
+            return response()->json([
+                'message' => 'geolocation information deleted.'
+            ]);
+        }
 
         return response()->json([
-            'message' => 'geolocation information deleted'
+            'success' => false,
+            'message' => 'geolocation information not found.'
         ]);
     }
 }
